feat: check scanner capabilities
This commit adds the `ScannerCapabilities` method used to
check the scanner capabilities.

Thanks to this method, we also perform a check of the capabilities
to determine if the user is going to use ADF Duplex mode on a
scanner that only supports ADF in Simplex mode.

This fixes #14.

The XML struct has been generated from my own scanner capabilities
and [this](https://github.com/alexpevzner/eSCL-protocol-traces/blob/master/Kyocera-ECOSYS-M2040dn.log)
log file.

On https://github.com/alexpevzner/eSCL-protocol-traces there are other
log files that we can use in case we want to extend the structs.

The XML Struct has been automatically generated using
https://github.com/twpayne/go-xmlstruct